@@ -4,11 +4,12 @@ require (
 	github.com/go-sql-driver/mysql v1.5.0
 	github.com/jmoiron/sqlx v1.2.0
 	github.com/lib/pq v1.9.0
-	github.com/mattn/go-sqlite3 v2.0.3+incompatible
+	github.com/mattn/go-sqlite3 v1.14.22
 	github.com/pkg/errors v0.9.1
 	golang.org/x/crypto v0.0.0-20201221181555-eec23a3978ad
 	golang.org/x/sys v0.0.0-20210113181707-4bcb84eeeb78
-	golang.org/x/term v0.0.0-20201210144234-2321bbc49cbf // indirect
 )
 
-go 1.13
+require golang.org/x/term v0.0.0-20201210144234-2321bbc49cbf // indirect
+
+go 1.22
